Do not delete `.keep` files when flushing cache

+++ src/RWFileCache.php
@@ -195,6 +195,9 @@ class RWFileCache
                     $result = rmdir($fullFilePath);
                 }
             } else {
+                if (basename($fullFilePath)=='.keep') {
+                    continue;
+                }
                 $result = unlink($fullFilePath);
             }
